+++ firmware/frame.h
@@ -4,5 +4,5 @@
 #include <stdbool.h>
 #include <stdint.h>
 
-extern uint16_t frame[WIDTH * HEIGHT];
+extern uint8_t frame[WIDTH * HEIGHT];
 extern bool frame_ready;
